pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2", "hatch-vcs>=0.3.0"]
build-backend = "hatchling.build"

[project]
name = "Byngosink"
authors = [{name = "Jamie", email = "jamiebloomfield8@gmail.com"}]
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "tornado==6.4.1",
    "shortuuid==1.0.13",
    "tortoise-orm==0.21.3",
    "orjson==3.10.5", # Accelerates tortoise JSON fields
    "pyjson5==1.6.6", # Allows commented json for use in generators.
]

[project.optional-dependencies]
dotenv = [
    'dotenv'
]